removido implementação com projeto reactor para supplir, add implementação mais limpar
