--- ds-pessoa/src/main/java/com/farias/banco/dspessoa/feignclients/ContaCorrenteFeignClients.java
@@ -1,10 +1,0 @@
-package com.farias.banco.dspessoa.feignclients;
-
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.stereotype.Component;
-
-@Component
-@FeignClient(name="ds-conta-corrente", path = "/contacorrente")
-public interface ContaCorrenteFeignClients {
-
-}
