--- ds-pessoa/src/main/java/com/farias/banco/dspessoa/handler/exception/DataBaseException.java
@@ -1,18 +1,0 @@
-package com.farias.banco.dspessoa.handler.exception;
-
-import org.springframework.http.HttpStatus;
-
-import lombok.Getter;
-
-@Getter
-public class DataBaseException extends RuntimeException {
-    private static final long serialVersionUID = 1L;
-
-    private final HttpStatus status;
-
-    public DataBaseException(HttpStatus status,  String message) {
-        super(message);
-        this.status = status;
-    }
-
-}
