+++ ds-pessoa/src/main/java/com/farias/banco/dspessoa/modules/handler/exception/DataBaseException.java
@@ -1,0 +1,18 @@
+package com.farias.banco.dspessoa.modules.handler.exception;
+
+import org.springframework.http.HttpStatus;
+
+import lombok.Getter;
+
+@Getter
+public class DataBaseException extends RuntimeException {
+    private static final long serialVersionUID = 1L;
+
+    private final HttpStatus status;
+
+    public DataBaseException(HttpStatus status,  String message) {
+        super(message);
+        this.status = status;
+    }
+
+}
