+++ ds-pessoa/src/main/java/com/farias/banco/dspessoa/modules/integration/feing/ContaCorrenteFeignClients.java
@@ -1,0 +1,10 @@
+package com.farias.banco.dspessoa.modules.integration.feing;
+
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.stereotype.Component;
+
+@Component
+@FeignClient(name="ds-conta-corrente", path = "/contacorrente")
+public interface ContaCorrenteFeignClients {
+
+}
