--- ds-pessoa/src/main/java/com/farias/banco/dspessoa/repository/PessoaRepository.java
@@ -1,11 +1,0 @@
-package com.farias.banco.dspessoa.repository;
-
-import com.farias.banco.dspessoa.model.Pessoa;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PessoaRepository extends JpaRepository<Pessoa, Long>, JpaSpecificationExecutor<Pessoa> {
-
-}
